Correcting some details to comply with repository rules
