+++ examples/trace_particles_coils_fullorbit_collisions.py
@@ -24,7 +24,7 @@ energy=4000*ONE_EV
 
 
 # Load coils and field
-json_file = os.path.join(os.path.dirname(__name__), 'input_files', 'ESSOS_biot_savart_LandremanPaulQA.json')
+json_file = os.path.join(os.path.dirname(__file__), 'input_files', 'ESSOS_biot_savart_LandremanPaulQA.json')
 coils = Coils_from_json(json_file)
 field = BiotSavart(coils)
